+++ src/Expression.ts
@@ -4,7 +4,6 @@
  */
 
 import Lexer from './Lexer.ts'
-import PromiseSync from './PromiseSync.ts'
 import Evaluator from './evaluator/Evaluator.ts'
 import Parser from './parser/Parser.ts'
 
@@ -14,8 +13,6 @@ interface Grammar {
   elements: Record<string, any>
   [key: string]: any
 }
-
-type PromiseConstructor = typeof Promise | typeof PromiseSync
 
 class Expression {
   _grammar: Grammar
@@ -44,44 +41,16 @@ class Expression {
   }
 
   /**
-   * Asynchronously evaluates the expression within an optional context.
-   * @param {Object} [context] A mapping of variables to values, which will be
-   *      made accessible to the Jexl expression when evaluating it
-   * @returns {Promise<*>} resolves with the result of the evaluation.
-   */
-  eval(context = {}) {
-    return this._eval(context, Promise)
-  }
-
-  /**
-   * Synchronously evaluates the expression within an optional context.
+   * Evaluates the expression within an optional context.
    * @param {Object} [context] A mapping of variables to values, which will be
    *      made accessible to the Jexl expression when evaluating it
    * @returns {*} the result of the evaluation.
    * @throws {*} on error
    */
-  evalSync(context = {}) {
-    const res = this._eval(context, PromiseSync as any) as PromiseSync
-    if (res.error) {
-      if (res.error instanceof Error) {
-        throw res.error
-      }
-      throw new Error(typeof res.error === 'string' ? res.error : JSON.stringify(res.error))
-    }
-    return res.value
-  }
-
-  _eval(context: any, promise: PromiseConstructor) {
-    return promise.resolve().then(() => {
-      const ast = this._getAst()
-      const evaluator = new Evaluator(
-        this._grammar,
-        context,
-        undefined,
-        promise
-      )
-      return evaluator.eval(ast!)
-    })
+  eval(context = {}) {
+    const ast = this._getAst()
+    const evaluator = new Evaluator(this._grammar, context)
+    return evaluator.eval(ast!)
   }
 
   _getAst() {