+++ src/Jexl.ts
@@ -9,7 +9,6 @@ import { getGrammar } from './grammar.ts'
 interface Grammar {
   elements: Record<string, any>
   functions: Record<string, (...args: any[]) => any>
-  transforms: Record<string, (val: any, ...args: any[]) => any>
 }
 
 /**
@@ -39,13 +38,11 @@ class Jexl {
    * @param {number} precedence The operator's precedence
    * @param {function} fn A function to run to calculate the result. The function
    *      will be called with two arguments: left and right, denoting the values
-   *      on either side of the operator. It should return either the resulting
-   *      value, or a Promise that resolves with the resulting value.
+   *      on either side of the operator. It should return the resulting value.
    * @param {boolean} [manualEval] If true, the `left` and `right` arguments
    *      will be wrapped in objects with an `eval` function. Calling
-   *      left.eval() or right.eval() will return a promise that resolves to
-   *      that operand's actual value. This is useful to conditionally evaluate
-   *      operands.
+   *      left.eval() or right.eval() will return that operand's actual value.
+   *      This is useful to conditionally evaluate operands.
    */
   addBinaryOp(
     operator: string,
@@ -88,8 +85,7 @@ class Jexl {
    * @param {string} operator The operator string to be added
    * @param {function} fn A function to run to calculate the result. The function
    *      will be called with one argument: the literal value to the right of the
-   *      operator. It should return either the resulting value, or a Promise
-   *      that resolves with the resulting value.
+   *      operator. It should return the resulting value.
    */
   addUnaryOp(operator: string, fn: (right: any) => any) {
     this._addGrammarElement(operator, {
@@ -97,28 +93,6 @@ class Jexl {
       weight: Infinity,
       eval: fn
     })
-  }
-
-  /**
-   * Adds or replaces a transform function in this Jexl instance.
-   * @param {string} name The name of the transform function, as it will be used
-   *      within Jexl expressions
-   * @param {function} fn The function to be executed when this transform is
-   *      invoked. It will be provided with at least one argument:
-   *          - {*} value: The value to be transformed
-   *          - {...*} args: The arguments for this transform
-   */
-  addTransform(name: string, fn: (val: any, ...args: any[]) => any) {
-    this._grammar.transforms[name] = fn
-  }
-
-  /**
-   * Syntactic sugar for calling {@link #addTransform} repeatedly.  This function
-   * accepts a map of one or more transform names to their transform function.
-   * @param {{}} map A map of transform names to transform functions
-   */
-  addTransforms(map: Record<string, (val: any, ...args: any[]) => any>) {
-    Object.assign(this._grammar.transforms, map)
   }
 
   /**
@@ -154,37 +128,16 @@ class Jexl {
   }
 
   /**
-   * Retrieves a previously set transform function.
-   * @param {string} name The name of the transform function
-   * @returns {function} The transform function
-   */
-  getTransform(name: string) {
-    return this._grammar.transforms[name]
-  }
-
-  /**
-   * Asynchronously evaluates a Jexl string within an optional context.
-   * @param {string} expression The Jexl expression to be evaluated
-   * @param {Object} [context] A mapping of variables to values, which will be
-   *      made accessible to the Jexl expression when evaluating it
-   * @returns {Promise<*>} resolves with the result of the evaluation.
-   */
-  eval(expression: string, context = {}) {
-    const exprObj = this.createExpression(expression)
-    return exprObj.eval(context)
-  }
-
-  /**
-   * Synchronously evaluates a Jexl string within an optional context.
+   * Evaluates a Jexl string within an optional context.
    * @param {string} expression The Jexl expression to be evaluated
    * @param {Object} [context] A mapping of variables to values, which will be
    *      made accessible to the Jexl expression when evaluating it
    * @returns {*} the result of the evaluation.
    * @throws {*} on error
    */
-  evalSync(expression: string, context = {}) {
+  eval(expression: string, context = {}) {
     const exprObj = this.createExpression(expression)
-    return exprObj.evalSync(context)
+    return exprObj.eval(context)
   }
 
   /**