+++ src/grammar.ts
@@ -9,7 +9,7 @@ export interface BinaryOp {
   type: 'binaryOp'
   precedence: number
   eval?: (left: any, right: any) => any
-  evalOnDemand?: (left: { eval: () => Promise<any> }, right: { eval: () => Promise<any> }) => Promise<any>
+  evalOnDemand?: (left: { eval: () => any }, right: { eval: () => any }) => any
 }
 
 export interface UnaryOp {
@@ -27,7 +27,6 @@ export type GrammarElement = BinaryOp | UnaryOp | SimpleElement
 export interface Grammar {
   elements: Record<string, GrammarElement>
   functions: Record<string, (...args: any[]) => any>
-  transforms: Record<string, (val: any, ...args: any[]) => any>
 }
 
 export const getGrammar = (): Grammar => ({
@@ -40,7 +39,6 @@ export const getGrammar = (): Grammar => ({
     '.': { type: 'dot' },
     '[': { type: 'openBracket' },
     ']': { type: 'closeBracket' },
-    '|': { type: 'pipe' },
     '{': { type: 'openCurl' },
     '}': { type: 'closeCurl' },
     ':': { type: 'colon' },
@@ -118,20 +116,22 @@ export const getGrammar = (): Grammar => ({
       type: 'binaryOp',
       precedence: 10,
       evalOnDemand: (left, right) => {
-        return left.eval().then((leftVal) => {
-          if (!leftVal) {return leftVal}
-          return right.eval()
-        })
+        const leftVal = left.eval()
+        if (!leftVal) {
+          return leftVal
+        }
+        return right.eval()
       }
     },
     '||': {
       type: 'binaryOp',
       precedence: 10,
       evalOnDemand: (left, right) => {
-        return left.eval().then((leftVal) => {
-          if (leftVal) {return leftVal}
-          return right.eval()
-        })
+        const leftVal = left.eval()
+        if (leftVal) {
+          return leftVal
+        }
+        return right.eval()
       }
     },
     in: {
@@ -169,32 +169,11 @@ export const getGrammar = (): Grammar => ({
    *       All of these are pre-evaluated to their actual values before calling
    *       the function.
    *
-   * The Jexl function should return either the transformed value, or
-   * a Promises/A+ Promise object that resolves with the value and rejects
-   * or throws only when an unrecoverable error occurs. Functions should
-   * generally return undefined when they don't make sense to be used on the
-   * given value type, rather than throw/reject. An error is only
-   * appropriate when the function would normally return a value, but
-   * cannot due to some other failure.
+   * The Jexl function should return the resulting value, or throw when an
+   * unrecoverable error occurs. Functions should generally return undefined
+   * when they don't make sense to be used on the given value type, rather
+   * than throw. An error is only appropriate when the function would normally
+   * return a value, but cannot due to some other failure.
    */
-  functions: {},
-
-  /**
-   * A map of transform names to transform functions. A transform function
-   * takes one ore more arguemnts:
-   *
-   *     - {*} val: A value to be transformed
-   *     - {*} ...args: A variable number of arguments passed to this transform.
-   *       All of these are pre-evaluated to their actual values before calling
-   *       the function.
-   *
-   * The transform function should return either the transformed value, or
-   * a Promises/A+ Promise object that resolves with the value and rejects
-   * or throws only when an unrecoverable error occurs. Transforms should
-   * generally return undefined when they don't make sense to be used on the
-   * given value type, rather than throw/reject. An error is only
-   * appropriate when the transform would normally return a value, but
-   * cannot due to some other failure.
-   */
-  transforms: {}
+  functions: {}
 })